@@ -4,8 +4,8 @@ version: 0.2.2
 homepage: https://github.com/letsar/overflow_view
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.20.0 <2.0.0"
+  sdk: ">=2.12.0-0.0 <3.0.0"
+  flutter: ">=1.24.0-10.2.pre <2.0.0"
 
 dependencies:
   flutter:
