@@ -22,10 +22,10 @@ class OverflowView extends MultiChildRenderObjectWidget {
   /// The [builder], [spacing] and [children] arguments must not be null.
   /// The [spacing] argument must also be positive and finite.
   OverflowView({
-    Key key,
-    @required OverflowIndicatorBuilder builder,
+    Key? key,
+    required OverflowIndicatorBuilder builder,
     Axis direction = Axis.horizontal,
-    @required List<Widget> children,
+    required List<Widget> children,
     double spacing = 0,
   }) : this._all(
           key: key,
@@ -43,10 +43,10 @@ class OverflowView extends MultiChildRenderObjectWidget {
   /// The [builder], [spacing] and [children] arguments must not be null.
   /// The [spacing] argument must also be positive and finite.
   OverflowView.flexible({
-    Key key,
-    @required OverflowIndicatorBuilder builder,
+    Key? key,
+    required OverflowIndicatorBuilder builder,
     Axis direction = Axis.horizontal,
-    @required List<Widget> children,
+    required List<Widget> children,
     double spacing = 0,
   }) : this._all(
           key: key,
@@ -58,17 +58,13 @@ class OverflowView extends MultiChildRenderObjectWidget {
         );
 
   OverflowView._all({
-    Key key,
-    @required OverflowIndicatorBuilder builder,
+    Key? key,
+    required OverflowIndicatorBuilder builder,
     this.direction = Axis.horizontal,
-    @required List<Widget> children,
+    required List<Widget> children,
     this.spacing = 0,
-    OverflowViewLayoutBehavior layoutBehavior,
-  })  : assert(builder != null),
-        assert(direction != null),
-        assert(children != null),
-        assert(spacing != null &&
-            spacing > double.negativeInfinity &&
+    required OverflowViewLayoutBehavior layoutBehavior,
+  })  : assert(spacing > double.negativeInfinity &&
             spacing < double.infinity),
         _layoutBehavior = layoutBehavior,
         super(
@@ -126,7 +122,7 @@ class _OverflowViewElement extends MultiChildRenderObjectElement {
   @override
   void debugVisitOnstageChildren(ElementVisitor visitor) {
     children.forEach((element) {
-      if (element.renderObject.isOnstage) {
+      if (element.renderObject?.isOnstage == true) {
         visitor(element);
       }
     });
